put all stuff in different folders. No errors anymore I think
