+++ Asterix/optical_systems.py
@@ -1,12 +1,9 @@
 # pylint: disable=invalid-name
 # pylint: disable=trailing-whitespace
 import os
-import inspect
 import copy
-import time
 import numpy as np
 from astropy.io import fits
-import skimage.transform
 
 import Asterix.processing_functions as proc
 import Asterix.propagation_functions as prop
@@ -674,1759 +671,3 @@ class OpticalSystem:
 
         return entrance_EF
 
-
-class Pupil(OpticalSystem):
-    """ --------------------------------------------------
-    initialize and describe the behavior of single pupil
-    pupil is a sub class of OpticalSystem.
-
-    Obviously you can define your pupil
-    without that with 2d arrray multiplication (this is a fairly simple object).
-
-    The main advantage of defining them using OpticalSystem is that you can
-    use default OpticalSystem functions to obtain PSF, transmission, etc...
-    and concatenate them with other elements
-
-    AUTHOR : Johan Mazoyer
-
-    
-    -------------------------------------------------- """
-
-    def __init__(self, modelconfig, prad=0., PupType=None, angle_rotation=0, Model_local_dir=None):
-        """ --------------------------------------------------
-        Initialize a pupil object.           
-        TODO: include an SCC Lyot pupil function here !
-        TODO: for now pupil .fits are monochromatic but the pupil propagation EF_through
-            use wavelenght as a parameter
-
-        AUTHOR : Johan Mazoyer
-
-        Parameters
-        ----------
-        modelconfig : dict
-                    general configuration parameters (sizes and dimensions)
-                        to initialize OpticalSystem class
-
-        prad : float
-            Default is the pupil radius in the parameter file (self.prad)
-            radius in pixels of the round pupil.
-
-        PupType : string (default None) 
-            Currently known possiibilities are 
-            "RoundPup", "ClearPlane", "RomanPup", "RomanLyot", "RomanPupTHD2", "RomanLyotTHD2"
-
-            if not one of those , it will try a full path to a fits file given by the user:
-
-            The pupil .fits files should be 2D and square with an even number of pix.
-            with even number of pix and centered between 4 pixels.
-            The array size will be assumed to correspond to the size of the entrance pupil 
-            and will be rescaled self.prad and then padded to self.dim_overpad_pupil
-
-            This is a bit dangerous because your .fits file might must be defined
-            the right way so be careful
-        
-        angle_rotation: float (default 0)
-            angle of rotation of the pupil in degrees in counter-clockwise direction.
-            this is only used if the pupil is not clear or empty
-
-
-        Model_local_dir: string, default None
-                    directory to save things you can measure yourself
-                    and can save to save time
-
-        -------------------------------------------------- """
-        # Initialize the OpticalSystem class and inherit properties
-        super().__init__(modelconfig)
-
-        if (Model_local_dir is not None) and not os.path.exists(Model_local_dir):
-            print("Creating directory " + Model_local_dir + " ...")
-            os.makedirs(Model_local_dir)
-
-        if prad == 0:
-            prad = self.prad
-
-        # known cases, with known responses
-        # default case: round pupil
-        if (PupType is None) or (PupType == "RoundPup"):
-            self.pup = phase_ampl.roundpupil(self.dim_overpad_pupil, prad)
-            angle_rotation = 0
-            if isinstance(prad, int) or prad.is_integer():
-                self.string_os += '_RoundPup' + str(int(prad))
-            else:
-                self.string_os += '_RoundPup' + str(round(prad, 1))
-
-        # Clear (in case we want to define an empty pupil plane)
-        elif PupType == "Clear":
-            self.pup = np.ones((self.dim_overpad_pupil, self.dim_overpad_pupil))
-            angle_rotation = 0
-            self.string_os += '_ClearPlane'
-
-        else:
-            # In those cases, we are using a fits to create the pupil
-            # in these first cases, we use a known .fits with hardcoded file name
-            if PupType == "RomanPup":
-                pup_fits = fits.getdata(os.path.join(model_dir, "roman_pup_500pix_center4pixels.fits"))
-                self.string_os += '_RomanPup' + str(int(prad))
-
-            elif PupType == "RomanPupTHD2":
-                pup_fits = fits.getdata(os.path.join(model_dir, "roman_pup_thd2_500pix_center4pixels.fits"))
-                self.string_os += '_RomanPupTHD2' + str(int(prad))
-
-            elif PupType == "RomanLyot":
-                pup_fits = fits.getdata(os.path.join(model_dir, "roman_lyot_500pix_center4pixels.fits"))
-                self.string_os += '_RomanLyot'
-
-            elif PupType == "RomanLyotTHD2":
-                pup_fits = fits.getdata(os.path.join(model_dir, "roman_lyot_thd2_500pix_center4pixels.fits"))
-                self.string_os += '_RomanLyotTHD2'
-
-            # finally in this last case, we use an unknown .fits defined by user
-            else:
-                if not os.path.exists(PupType):
-
-                    print("""
-                            filename_instr_XXX parameters must either be a known keyword 
-                            'RoundPup', 'Clear', 'RomanPup', 'RomanLyot' , 'RomanPupTHD2', 'THD2',
-                            or an exisiting full path .fits name. This is not the case here,
-                            the name  '{0}' is not a known keyword and is not an existing filename
-                            """.format(PupType))
-                    print("")
-                    print("")
-                    raise
-
-                # this is an existing fits file
-                # we start by a bunch of tests to check
-                # that pupil has a certain acceptable form.
-                # print("we load the pupil: " + filename)
-                # print("we assume it is centered in its array")
-                pup_fits = fits.getdata(PupType)
-
-                if len(pup_fits.shape) != 2:
-                    raise Exception("file " + PupType + " should be a 2D array")
-
-                if pup_fits.shape[0] != pup_fits.shape[1]:
-                    raise Exception("file " + PupType + " appears to be not square")
-
-                self.string_os += '_Fits'
-
-            if angle_rotation != 0:
-                pup_fits = skimage.transform.rotate(pup_fits, angle_rotation, preserve_range=True)
-                self.string_os += 'Rot' + str(int(angle_rotation))
-
-                fits.writeto(os.path.join(Model_local_dir,
-                                          PupType + 'Rot' + str(int(angle_rotation)) + '.fits'),
-                             pup_fits,
-                             overwrite=True)
-
-            # we have the fits, we now rescale to good size
-            if pup_fits.shape[0] == 2 * self.prad:
-                pup_fits_right_size = pup_fits
-            else:
-                #Rescale to the pupil size
-                find_divisors = list()
-                for i in range(60, pup_fits.shape[0] + 1):
-                    if pup_fits.shape[0] % i == 0:
-                        find_divisors.append(i)
-
-                if not int(2 * self.prad) in find_divisors:
-                    raise Exception(
-                        "Choose a divisor of the .fits file size ({0}) for diam_pup_in_pix parameter: {1}".
-                        format(pup_fits.shape[0], find_divisors))
-                else:
-                    pup_fits_right_size = proc.rebin(pup_fits,
-                                                     int(pup_fits.shape[0] / (2 * self.prad)),
-                                                     center_on_pixel=False)
-
-            self.pup = proc.crop_or_pad_image(pup_fits_right_size, self.dim_overpad_pupil)
-
-        #initialize the max and sum of PSFs for the normalization to contrast
-        self.measure_normalization()
-
-    def EF_through(self,
-                   entrance_EF=1.,
-                   wavelength=None,
-                   save_all_planes_to_fits=False,
-                   dir_save_all_planes=None,
-                   **kwargs):
-        """ --------------------------------------------------
-        Propagate the electric field through the pupil
-        AUTHOR : Johan Mazoyer
-
-        Parameters
-        ----------
-        entrance_EF:    2D complex array of size [self.dim_overpad_pupil, self.dim_overpad_pupil]
-                        Can also be a float scalar in which case entrance_EF is constant
-                        default is 1.
-                        Electric field in the pupil plane a the entrance of the system.
-
-        wavelength : float. Default is self.wavelength_0 the reference wavelength
-                current wavelength in m.
-
-        save_all_planes_to_fits: Bool, default False.
-                if True, save all planes to fits for debugging purposes to dir_save_all_planes
-                This can generate a lot of fits especially if in a loop so the code force you
-                to define a repository.
-
-        dir_save_all_planes : default None. 
-                                directory to save all plane
-                                in fits if save_all_planes_to_fits = True
-
-        Returns
-        ------
-        exit_EF : 2D array, of size [self.dim_overpad_pupil, self.dim_overpad_pupil]
-                        Electric field in the pupil plane a the exit of the system
-
-
-        -------------------------------------------------- """
-
-        # call the OpticalSystem super function to check and format the variable entrance_EF
-        entrance_EF = super().EF_through(entrance_EF=entrance_EF)
-        if wavelength is None:
-            wavelength = self.wavelength_0
-
-        if save_all_planes_to_fits == True:
-            name_plane = 'EF_PP_before_pupil' + '_wl{}'.format(int(wavelength * 1e9))
-            useful.save_plane_in_fits(dir_save_all_planes, name_plane, entrance_EF)
-
-        if len(self.pup.shape) == 2:
-            exit_EF = entrance_EF * self.pup
-
-        elif len(self.pup.shape) == 3:
-            if self.pup.shape != self.nb_wav:
-                raise Exception("I'm confused, your pupil seem to be polychromatic" +
-                                "(pup.shape=3) but the # of WL (pup.shape[0]={}) ".format(self.pup.shape[0]) +
-                                "is different from the system # of WL (nb_wav={})".format(self.nb_wav))
-            else:
-                exit_EF = entrance_EF * self.pup[self.wav_vec.tolist().index(wavelength)]
-        else:
-            raise Exception("pupil dimension are not acceptable")
-
-        if save_all_planes_to_fits == True:
-            name_plane = 'EF_PP_after_pupil' + '_wl{}'.format(int(wavelength * 1e9))
-            useful.save_plane_in_fits(dir_save_all_planes, name_plane, exit_EF)
-
-        return exit_EF
-
-
-class Coronagraph(OpticalSystem):
-    """ --------------------------------------------------
-    initialize and describe the behavior of a coronagraph system (from apod plane to the Lyot plane)
-    coronagraph is a sub class of OpticalSystem.
-
-    AUTHOR : Johan Mazoyer
-
-    -------------------------------------------------- """
-
-    def __init__(self, modelconfig, coroconfig, Model_local_dir=None):
-        """ --------------------------------------------------
-        Initialize a coronograph object
-
-        AUTHOR : Johan Mazoyer
-
-        Parameters
-        ----------
-        modelconfig : dict
-                general configuration parameters (sizes and dimensions)
-        coroconfig : : dict
-                coronagraph parameters
-
-        Model_local_dir: string, default None
-                    directory to save things you can measure yourself
-                    and can save to save times      
-        
-
-        -------------------------------------------------- """
-
-        # Initialize the OpticalSystem class and inherit properties
-        super().__init__(modelconfig)
-
-        if (Model_local_dir is not None) and not os.path.exists(Model_local_dir):
-            print("Creating directory " + Model_local_dir + " ...")
-            os.makedirs(Model_local_dir)
-
-        # Plane at the entrance of the coronagraph. In THD2, this is an empty plane.
-        # In Roman this is where is the apodiser
-        self.apod_pup = Pupil(modelconfig,
-                              prad=self.prad,
-                              PupType=coroconfig["filename_instr_apod"],
-                              angle_rotation=coroconfig['apod_pup_rotation'],
-                              Model_local_dir=Model_local_dir)
-
-        self.string_os += '_Apod' + self.apod_pup.string_os
-
-        #coronagraph focal plane mask type
-        self.corona_type = coroconfig["corona_type"].lower()
-
-        self.string_os += '_' + self.corona_type
-
-        # dim_fp_fft definition only use if prop_apod2lyot == 'fft'
-        self.corono_fpm_sampling = self.Science_sampling
-        self.dim_fp_fft = np.zeros(len(self.wav_vec), dtype=np.int)
-        for i, wav in enumerate(self.wav_vec):
-            self.dim_fp_fft[i] = int(np.ceil(
-                self.prad * self.corono_fpm_sampling * self.wavelength_0 / wav)) * 2
-            # we take the ceil to be sure that we measure at least the good resolution
-            # We do not need to be exact, the mft in science_focal_plane will be
-
-        if self.corona_type == "fqpm":
-            self.prop_apod2lyot = 'mft'
-            self.err_fqpm = coroconfig["err_fqpm"]
-            self.achrom_fqpm = coroconfig["achrom_fqpm"]
-            self.FPmsk = self.FQPM()
-            if self.achrom_fqpm:
-                str_achrom = "achrom"
-            else:
-                str_achrom = "nonachrom"
-            self.string_os += '_' + str_achrom
-            self.perfect_coro = True
-
-        elif self.corona_type == "classiclyot" or self.corona_type == "hlc":
-            self.prop_apod2lyot = 'mft-babinet'
-            self.rad_lyot_fpm = coroconfig["rad_lyot_fpm"]
-            # we oversample the center in babinet's mode because we can
-            # hard coded for now, this is very internal cooking
-
-            self.Lyot_fpm_sampling = 20.  #self.Science_sampling
-            rad_LyotFP_pix = self.rad_lyot_fpm * self.Lyot_fpm_sampling
-            self.dim_fpm = 2 * int(2.2 * rad_LyotFP_pix / 2)
-
-            self.string_os += '_' + "iwa" + str(round(self.rad_lyot_fpm, 2))
-            self.perfect_coro = False
-            if self.corona_type == "classiclyot":
-                self.FPmsk = self.ClassicalLyot()
-            else:
-                self.transmission_fpm = coroconfig["transmission_fpm"]
-                self.phase_fpm = coroconfig["phase_fpm"]
-                self.string_os += '_' + "trans{:.1e}".format(self.transmission_fpm) + "_pha{0}".format(
-                    round(self.phase_fpm, 2))
-                self.FPmsk = self.HLC()
-
-        elif self.corona_type == "knife":
-            self.prop_apod2lyot = 'mft'
-            self.coro_position = coroconfig["knife_coro_position"].lower()
-            self.knife_coro_offset = coroconfig["knife_coro_offset"]
-            self.FPmsk = self.KnifeEdgeCoro()
-            self.string_os += '_' + self.coro_position + "_iwa" + str(round(self.knife_coro_offset, 2))
-            self.perfect_coro = False
-
-        elif self.corona_type == "vortex":
-            self.prop_apod2lyot = 'mft'
-            vortex_charge = coroconfig["vortex_charge"]
-            self.string_os += '_charge' + str(int(vortex_charge))
-            self.FPmsk = self.Vortex(vortex_charge=vortex_charge)
-            self.perfect_coro = True
-
-        elif self.corona_type == "wrapped_vortex":
-            self.prop_apod2lyot = 'mft'
-            self.string_os += '2020'
-            self.FPmsk = list([
-                self.EF_from_phase_and_ampl(phase_abb=proc.crop_or_pad_image(
-                    fits.getdata(model_dir + coroconfig["wrapped_vortex_fits_file"]), self.dimScience))
-            ])
-            self.perfect_coro = True
-
-        else:
-            raise Exception(f"The requested coronagraph mode '{self.corona_type}' does not exists.")
-
-        self.lyot_pup = Pupil(modelconfig,
-                              prad=self.prad * coroconfig["diam_lyot_in_m"] / self.diam_pup_in_m,
-                              PupType=coroconfig["filename_instr_lyot"],
-                              angle_rotation=coroconfig['lyot_pup_rotation'],
-                              Model_local_dir=Model_local_dir)
-
-        self.string_os += '_LS' + self.lyot_pup.string_os
-
-        if "bool_overwrite_perfect_coro" in coroconfig:
-            if coroconfig["bool_overwrite_perfect_coro"] is True:
-                self.perfect_coro = True
-            if coroconfig["bool_overwrite_perfect_coro"] is False:
-                self.perfect_coro = False
-
-        if self.perfect_coro is True:
-
-            if coroconfig["filename_instr_apod"] == "Clear":
-                # We need a round pupil only to measure the response
-                # of the coronograph to a round pupil to remove it
-                # THIS IS NOT THE ENTRANCE PUPIL,
-                # this is a round pupil of the same size
-                pup_for_perfect_coro = Pupil(modelconfig, prad=self.prad)
-
-                # do a propagation once with self.perfect_Lyot_pupil = 0 to
-                # measure the Lyot pupil that will be removed after
-                self.perfect_Lyot_pupil = [0] * self.nb_wav
-                for i, wave_here in enumerate(self.wav_vec):
-                    self.perfect_Lyot_pupil[i] = self.EF_through(
-                        entrance_EF=pup_for_perfect_coro.EF_through(wavelength=wave_here),
-                        wavelength=wave_here)
-            else:
-                # In this case we have a coronagraph entrance pupil.
-                # Do a propagation once with self.perfect_Lyot_pupil = 0 to
-                # measure the Lyot pupil that will be removed after.
-                self.perfect_Lyot_pupil = [0] * self.nb_wav
-                for i, wave_here in enumerate(self.wav_vec):
-                    self.perfect_Lyot_pupil[i] = self.EF_through(wavelength=wave_here)
-
-        # Initialize the max and sum of PSFs for the normalization to contrast
-        self.measure_normalization()
-
-    def EF_through(self,
-                   entrance_EF=1.,
-                   wavelength=None,
-                   noFPM=False,
-                   EF_aberrations_introduced_in_LS=1.,
-                   save_all_planes_to_fits=False,
-                   dir_save_all_planes=None,
-                   **kwargs):
-        """ --------------------------------------------------
-        Propagate the electric field from apod plane before the apod
-        pupil to Lyot plane after Lyot pupil
-
-        AUTHOR : Johan Mazoyer
-
-        03/22 : Correction in the babinet propagation
-
-        Parameters
-        ----------
-        entrance_EF:    2D complex array of size [self.dim_overpad_pupil, self.dim_overpad_pupil]
-                        Can also be a float scalar in which case entrance_EF is constant
-                        default is 1.
-                        Electric field in the pupil plane a the entrance of the system.
-
-        wavelength : float. Default is self.wavelength_0 the reference wavelength
-                        current wavelength in m.
-
-        noFPM : bool (default: False)
-            if True, remove the FPM if one want to measure a un-obstructed PSF
-        
-        EF_aberrations_introduced_in_LS: 2D complex array of size [self.dim_overpad_pupil, self.dim_overpad_pupil]
-                        Can also be a float scalar in which case entrance_EF is constant
-                        default is 1.
-                        electrical field created by the downstream aberrations introduced directly in the Lyot Stop
-
-        save_all_planes_to_fits: Bool, default False.
-                if True, save all planes to fits for debugging purposes to dir_save_all_planes
-                This can generate a lot of fits especially if in a loop so the code force you
-                to define a repository.
-        
-        dir_save_all_planes : default None. 
-                                directory to save all plane in
-                              fits if save_all_planes_to_fits = True
-
-        Returns
-        ------
-        exit_EF : 2D array, of size [self.dim_overpad_pupil, self.dim_overpad_pupil]
-                Electric field in the pupil plane a the exit of the system
-        
-
-        -------------------------------------------------- """
-
-        # call the OpticalSystem super function to check and format the variable entrance_EF
-        entrance_EF = super().EF_through(entrance_EF=entrance_EF)
-
-        if wavelength is None:
-            wavelength = self.wavelength_0
-
-        if save_all_planes_to_fits == True:
-            name_plane = 'EF_PP_before_apod' + '_wl{}'.format(int(wavelength * 1e9))
-            useful.save_plane_in_fits(dir_save_all_planes, name_plane, entrance_EF)
-
-        if noFPM:
-            FPmsk = 1.
-        else:
-            FPmsk = self.FPmsk[self.wav_vec.tolist().index(wavelength)]
-
-        lambda_ratio = wavelength / self.wavelength_0
-
-        input_wavefront_after_apod = self.apod_pup.EF_through(entrance_EF=entrance_EF, wavelength=wavelength)
-
-        if save_all_planes_to_fits == True:
-            name_plane = 'apod' + '_wl{}'.format(int(wavelength * 1e9))
-            useful.save_plane_in_fits(dir_save_all_planes, name_plane, self.apod_pup.pup)
-
-            name_plane = 'EF_PP_after_apod' + '_wl{}'.format(int(wavelength * 1e9))
-            useful.save_plane_in_fits(dir_save_all_planes, name_plane, input_wavefront_after_apod)
-
-        # we take the convention that for all propation methods, the PSF must be
-        # "in between 4 pixels" in the focal plane.
-
-        if self.prop_apod2lyot == "fft":
-            dim_fp_fft_here = self.dim_fp_fft[self.wav_vec.tolist().index(wavelength)]
-            input_wavefront_after_apod_pad = proc.crop_or_pad_image(input_wavefront_after_apod,
-                                                                    dim_fp_fft_here)
-
-            corono_focal_plane = prop.fft_choosecenter(input_wavefront_after_apod_pad,
-                                                       inverse=False,
-                                                       center_pos='bb',
-                                                       norm='ortho')
-
-            if save_all_planes_to_fits == True:
-                name_plane = 'EF_FP_before_FPM' + '_wl{}'.format(int(wavelength * 1e9))
-                useful.save_plane_in_fits(dir_save_all_planes, name_plane,
-                                          np.fft.fftshift(corono_focal_plane))
-
-                name_plane = 'PSF EF_FP_before_FPM' + '_wl{}'.format(int(wavelength * 1e9))
-                useful.save_plane_in_fits(dir_save_all_planes, name_plane,
-                                          np.fft.fftshift(np.abs(corono_focal_plane)**2))
-                if not noFPM:
-                    name_plane = 'FPM' + '_wl{}'.format(int(wavelength * 1e9))
-                    useful.save_plane_in_fits(dir_save_all_planes, name_plane, FPmsk)
-
-                    name_plane = 'FPMphase' + '_wl{}'.format(int(wavelength * 1e9))
-                    useful.save_plane_in_fits(dir_save_all_planes, name_plane, np.angle(FPmsk))
-
-                    name_plane = 'FPMmod' + '_wl{}'.format(int(wavelength * 1e9))
-                    useful.save_plane_in_fits(dir_save_all_planes, name_plane, np.abs(FPmsk))
-
-                name_plane = 'EF_FP_after_FPM' + '_wl{}'.format(int(wavelength * 1e9))
-                useful.save_plane_in_fits(dir_save_all_planes, name_plane,
-                                          np.fft.fftshift(corono_focal_plane * FPmsk))
-
-            # Focal plane to Lyot plane
-            lyotplane_before_lyot = prop.fft_choosecenter(corono_focal_plane * FPmsk,
-                                                          inverse=True,
-                                                          center_pos='bb',
-                                                          norm='ortho')
-
-        elif self.prop_apod2lyot == "mft-babinet":
-            #Apod plane to focal plane
-
-            corono_focal_plane = prop.mft(input_wavefront_after_apod,
-                                          int(2 * self.prad),
-                                          self.dim_fpm,
-                                          self.dim_fpm / self.Lyot_fpm_sampling * lambda_ratio,
-                                          inverse=False,
-                                          norm='ortho')
-
-            if save_all_planes_to_fits == True:
-                name_plane = 'EF_FP_before_FPM' + '_wl{}'.format(int(wavelength * 1e9))
-                useful.save_plane_in_fits(dir_save_all_planes, name_plane, corono_focal_plane)
-                if not noFPM:
-                    name_plane = 'FPM' + '_wl{}'.format(int(wavelength * 1e9))
-                    useful.save_plane_in_fits(dir_save_all_planes, name_plane, FPmsk)
-
-                    name_plane = 'FPMphase' + '_wl{}'.format(int(wavelength * 1e9))
-                    useful.save_plane_in_fits(dir_save_all_planes, name_plane, np.angle(FPmsk))
-
-                    name_plane = 'FPMmod' + '_wl{}'.format(int(wavelength * 1e9))
-                    useful.save_plane_in_fits(dir_save_all_planes, name_plane, np.abs(FPmsk))
-
-                name_plane = 'EF_FP_after_FPM' + '_wl{}'.format(int(wavelength * 1e9))
-                useful.save_plane_in_fits(dir_save_all_planes, name_plane, corono_focal_plane * FPmsk)
-
-                name_plane = 'EF_FP_after_1minusFPM' + '_wl{}'.format(int(wavelength * 1e9))
-                useful.save_plane_in_fits(dir_save_all_planes, name_plane, corono_focal_plane * (1 - FPmsk))
-
-            # Focal plane to Lyot plane
-            # Babinet's trick:
-            lyotplane_before_lyot_central_part = proc.crop_or_pad_image(
-                prop.mft(corono_focal_plane * (1 - FPmsk),
-                         self.dim_fpm,
-                         int(2 * self.prad),
-                         self.dim_fpm / self.Lyot_fpm_sampling * lambda_ratio,
-                         inverse=True,
-                         norm='ortho'), self.dim_overpad_pupil)
-
-            # Babinet's trick
-            lyotplane_before_lyot = input_wavefront_after_apod - lyotplane_before_lyot_central_part
-
-        elif self.prop_apod2lyot == "mft":
-            # Apod plane to focal plane
-
-            corono_focal_plane = prop.mft(input_wavefront_after_apod,
-                                          int(2 * self.prad),
-                                          self.dimScience,
-                                          self.dimScience / self.Science_sampling * lambda_ratio,
-                                          inverse=False,
-                                          norm='ortho')
-
-            if save_all_planes_to_fits == True:
-                name_plane = 'EF_FP_before_FPM' + '_wl{}'.format(int(wavelength * 1e9))
-                useful.save_plane_in_fits(dir_save_all_planes, name_plane, corono_focal_plane)
-                if not noFPM:
-                    name_plane = 'FPM' + '_wl{}'.format(int(wavelength * 1e9))
-                    useful.save_plane_in_fits(dir_save_all_planes, name_plane, FPmsk)
-
-                    name_plane = 'FPMphase' + '_wl{}'.format(int(wavelength * 1e9))
-                    useful.save_plane_in_fits(dir_save_all_planes, name_plane, np.angle(FPmsk))
-
-                    name_plane = 'FPMmod' + '_wl{}'.format(int(wavelength * 1e9))
-                    useful.save_plane_in_fits(dir_save_all_planes, name_plane, np.abs(FPmsk))
-
-                name_plane = 'EF_FP_after_FPM' + '_wl{}'.format(int(wavelength * 1e9))
-                useful.save_plane_in_fits(dir_save_all_planes, name_plane, corono_focal_plane * FPmsk)
-
-            # Focal plane to Lyot plane
-            lyotplane_before_lyot = proc.crop_or_pad_image(
-                prop.mft(corono_focal_plane * FPmsk,
-                         self.dimScience,
-                         int(2 * self.prad),
-                         self.dimScience / self.Science_sampling * lambda_ratio,
-                         inverse=True,
-                         norm='ortho'), self.dim_overpad_pupil)
-
-        else:
-            raise Exception(self.prop_apod2lyot + " is not a known prop_apod2lyot propagation mehtod")
-
-        if save_all_planes_to_fits == True:
-            name_plane = 'EF_PP_before_LS' + '_wl{}'.format(int(wavelength * 1e9))
-            useful.save_plane_in_fits(dir_save_all_planes, name_plane, lyotplane_before_lyot)
-
-        # we add the downstream aberrations if we need them
-        lyotplane_before_lyot *= EF_aberrations_introduced_in_LS
-
-        # crop to the dim_overpad_pupil expeted size
-        lyotplane_before_lyot_crop = proc.crop_or_pad_image(lyotplane_before_lyot, self.dim_overpad_pupil)
-
-        # Field after filtering by Lyot stop
-        lyotplane_after_lyot = self.lyot_pup.EF_through(entrance_EF=lyotplane_before_lyot_crop,
-                                                        wavelength=wavelength)
-
-        if (self.perfect_coro == True) & (noFPM == False):
-            lyotplane_after_lyot = lyotplane_after_lyot - self.perfect_Lyot_pupil[self.wav_vec.tolist().index(
-                wavelength)]
-
-        if save_all_planes_to_fits == True:
-            name_plane = 'LS' + '_wl{}'.format(int(wavelength * 1e9))
-            useful.save_plane_in_fits(dir_save_all_planes, name_plane, self.lyot_pup.pup)
-
-            name_plane = 'EF_PP_after_LS' + '_wl{}'.format(int(wavelength * 1e9))
-            useful.save_plane_in_fits(dir_save_all_planes, name_plane, lyotplane_after_lyot)
-
-        return lyotplane_after_lyot
-
-    def FQPM(self):
-        """ --------------------------------------------------
-        Create a Four Quadrant Phase Mask coronagraph
-        AUTHOR : Axel Potier
-        Modified by Johan Mazoyer
-
-
-        Returns
-        ------
-        FQPM : list of len(self.wav_vec) 2D arrays   
-            complex transmission of the FQPM mask at all wl
-        
-        
-        -------------------------------------------------- """
-
-        if self.prop_apod2lyot == "fft":
-            maxdimension_array_fpm = np.max(self.dim_fp_fft)
-        else:
-            maxdimension_array_fpm = self.dimScience
-
-        xx, yy = np.meshgrid(
-            np.arange(maxdimension_array_fpm) - (maxdimension_array_fpm) / 2,
-            np.arange(maxdimension_array_fpm) - (maxdimension_array_fpm) / 2)
-
-        fqpm_thick_vert = np.zeros((maxdimension_array_fpm, maxdimension_array_fpm))
-        fqpm_thick_vert[np.where(xx < 0)] = 1
-        fqpm_thick_hor = np.zeros((maxdimension_array_fpm, maxdimension_array_fpm))
-        fqpm_thick_hor[np.where(yy >= 0)] = 1
-        fqpm_thick = fqpm_thick_vert - fqpm_thick_hor
-
-        fqpm = list()
-        for i, wav in enumerate(self.wav_vec):
-            if self.prop_apod2lyot == "fft":
-                dim_fp = self.dim_fp_fft[i]
-            else:
-                dim_fp = self.dimScience
-
-            phase4q = np.zeros((dim_fp, dim_fp))
-            fqpm_thick_cut = proc.crop_or_pad_image(fqpm_thick, dim_fp)
-            phase4q[np.where(fqpm_thick_cut != 0)] = (np.pi + self.err_fqpm)
-
-            if self.achrom_fqpm == True:
-                # if we want to do an an achromatic_fqpm, we do not include a variation
-                # of the phase with the wl.
-                fqpm.append(np.exp(1j * phase4q))
-            else:
-                # in the general case, we use the EF_from_phase_and_ampl which handle the phase
-                # chromaticity.
-                fqpm.append(self.EF_from_phase_and_ampl(phase_abb=phase4q, wavelengths=wav))
-
-        return fqpm
-
-    def Vortex(self, vortex_charge=2):
-        """ --------------------------------------------------
-        Create a charge2 vortex.
-
-        AUTHOR : Johan Mazoyer
-
-        Parameters
-        ------
-        Charge : int, defaut 2
-            charge of the vortex. can be 2, 4, 6. Defaut is charge 2
-
-        Returns
-        ------
-        vortex_fpm : list of 2D numpy array
-                            the FP mask at all wl
-
-
-        -------------------------------------------------- """
-
-        if self.prop_apod2lyot == "fft":
-            maxdimension_array_fpm = np.max(self.dim_fp_fft)
-        else:
-            maxdimension_array_fpm = self.dimScience
-
-        xx, yy = np.meshgrid(
-            np.arange(maxdimension_array_fpm) - (maxdimension_array_fpm) / 2 + 1 / 2,
-            np.arange(maxdimension_array_fpm) - (maxdimension_array_fpm) / 2 + 1 / 2)
-
-        phase_vortex = vortex_charge * np.angle(xx + 1j * yy)
-
-        vortex = list()
-        for i, wav in enumerate(self.wav_vec):
-            if self.prop_apod2lyot == "fft":
-                dim_fp = self.dim_fp_fft[i]
-            else:
-                dim_fp = self.dimScience
-
-            phasevortex_cut = proc.crop_or_pad_image(phase_vortex,
-                                                     dim_fp)  #*phase_ampl.roundpupil(dim_fp, dim_fp/2)
-            vortex.append(np.exp(1j * phasevortex_cut))
-
-        return vortex
-
-    def KnifeEdgeCoro(self):
-        """ --------------------------------------------------
-        Create a Knife edge coronagraph of size (dimScience,dimScience)
-        AUTHOR : Axel Potier
-        Modified by Johan Mazoyer
-
-        Returns
-        ------
-        Knife FPM : list of len(self.wav_vec) 2D arrays 
-                    gcomplex transmission of the Knife edge coronagraph mask at all wl
-
-
-        -------------------------------------------------- """
-        if self.prop_apod2lyot == "fft":
-            maxdimension_array_fpm = np.max(self.dim_fp_fft)
-            if len(self.wav_vec) > 1:
-                raise Exception("knife currently not coded in polychromatic fft")
-        else:
-            maxdimension_array_fpm = self.dimScience
-
-        # self.coro_position can be 'left', 'right', 'top' or 'bottom'
-        # to define the orientation of the coronagraph
-
-        #  Number of pixels per resolution element at central wavelength
-
-        xx, yy = np.meshgrid(np.arange(maxdimension_array_fpm), np.arange(maxdimension_array_fpm))
-
-        Knife = np.zeros((maxdimension_array_fpm, maxdimension_array_fpm))
-        if self.coro_position == "right":
-            Knife[np.where(
-                xx > (maxdimension_array_fpm / 2 + self.knife_coro_offset * self.Science_sampling))] = 1
-        if self.coro_position == "left":
-            Knife[np.where(
-                xx < (maxdimension_array_fpm / 2 - self.knife_coro_offset * self.Science_sampling))] = 1
-        if self.coro_position == "bottom":
-            Knife[np.where(
-                yy > (maxdimension_array_fpm / 2 + self.knife_coro_offset * self.Science_sampling))] = 1
-        if self.coro_position == "top":
-            Knife[np.where(
-                yy < (maxdimension_array_fpm / 2 - self.knife_coro_offset * self.Science_sampling))] = 1
-
-        knife_allwl = list()
-        for i in range(len(self.wav_vec)):
-            knife_allwl.append(Knife)
-
-        return knife_allwl
-
-    def ClassicalLyot(self):
-        """ --------------------------------------------------
-        Create a classical Lyot coronagraph of radius rad_LyotFP 0
-        AUTHOR : Johan Mazoyer
-
-        Returns
-        ------
-        classical Lyot fpm : list of 2D numpy array
-                            the FP mask at all wl
-
-        
-        -------------------------------------------------- """
-
-        rad_LyotFP_pix = self.rad_lyot_fpm * self.Lyot_fpm_sampling
-
-        ClassicalLyotFPM = 1. - phase_ampl.roundpupil(self.dim_fpm, rad_LyotFP_pix)
-
-        ClassicalLyotFPM_allwl = list()
-        for wav in self.wav_vec:
-            ClassicalLyotFPM_allwl.append(ClassicalLyotFPM)
-
-        return ClassicalLyotFPM_allwl
-
-    def HLC(self):
-        """ --------------------------------------------------
-        Create a HLC of radius rad_LyotFP 0
-        AUTHOR : Johan Mazoyer
-
-        Returns
-        ------
-        classical Lyot hlc : list of 2D numpy array
-                            the FP mask at all wl
-        
-        
-        -------------------------------------------------- """
-
-        # we create a Classical Lyot Focal plane
-        ClassicalLyotFP = self.ClassicalLyot()[0]
-
-        whClassicalLyotstop = np.where(ClassicalLyotFP == 0.)
-
-        # we define phase and amplitude for the HLC at the reference WL
-        phase_hlc = np.zeros(ClassicalLyotFP.shape)
-        phase_hlc[whClassicalLyotstop] = self.phase_fpm
-        # transmission_fpm is defined in intensity and EF_from_phase_and_ampl takes amplitude
-        ampl_hlc = np.zeros(ClassicalLyotFP.shape)
-        ampl_hlc[whClassicalLyotstop] = np.sqrt(self.transmission_fpm) - 1
-
-        hlc_all_wl = list()
-        for wav in self.wav_vec:
-            hlc_all_wl.append(
-                self.EF_from_phase_and_ampl(ampl_abb=ampl_hlc, phase_abb=phase_hlc, wavelengths=wav))
-
-        return hlc_all_wl
-
-
-class DeformableMirror(OpticalSystem):
-    """ --------------------------------------------------
-    initialize and describe the behavior of a deformable mirror
-    (in pupil plane or out of pupil plane)
-    coronagraph is a sub class of OpticalSystem.
-
-
-    AUTHOR : Johan Mazoyer
-
-
-    -------------------------------------------------- """
-
-    def __init__(self, modelconfig, DMconfig, Name_DM='DM3', Model_local_dir=None):
-        """ --------------------------------------------------
-        Initialize a deformable mirror object
-
-        AUTHOR : Johan Mazoyer
-
-        Parameters
-        ----------
-        modelconfig : dict
-                general configuration parameters (sizes and dimensions)
-
-        DMconfig : dict
-                DM configuration parameters dictionary 
-
-        Name_DM : string
-                the name of the DM, which allows to find it in the parameter file
-                we measure and save the pushact functions
-
-        Model_local_dir: path
-                directory to save things you can measure yourself
-                    and can save to save time
-        
-        
-        -------------------------------------------------- """
-
-        # Initialize the OpticalSystem class and inherit properties
-        super().__init__(modelconfig)
-
-        if not os.path.exists(Model_local_dir):
-            print("Creating directory " + Model_local_dir + " ...")
-            os.makedirs(Model_local_dir)
-
-        self.Model_local_dir = Model_local_dir
-
-        self.Name_DM = Name_DM
-        self.z_position = DMconfig[self.Name_DM + "_z_position"]
-        self.active = DMconfig[self.Name_DM + "_active"]
-
-        self.WhichInPup_threshold = DMconfig["MinimumSurfaceRatioInThePupil"]
-
-        # For intialization, we assume no misregistration, we introduce it after
-        # estimation and correction matrices are created.
-        self.misregistration = False
-
-        if DMconfig[self.Name_DM + "_Generic"] == True:
-            self.total_act = DMconfig[self.Name_DM + "_Nact1D"]**2
-            self.number_act = self.total_act
-            self.active_actuators = np.arange(self.number_act)
-        else:
-            # first thing we do is to open filename_grid_actu to check the number of
-            # actuator of this DM. We need the number of act to read and load pushact .fits
-            self.total_act = fits.getdata(model_dir + DMconfig[self.Name_DM + "_filename_grid_actu"]).shape[1]
-
-            if DMconfig[self.Name_DM + "_filename_active_actu"] != "":
-                self.active_actuators = fits.getdata(model_dir +
-                                                     DMconfig[self.Name_DM +
-                                                              "_filename_active_actu"]).astype(int)
-                self.number_act = len(self.active_actuators)
-
-            else:
-                self.number_act = self.total_act
-                self.active_actuators = np.arange(self.number_act)
-
-        self.string_os += '_' + self.Name_DM + "_z" + str(int(self.z_position * 1000)) + "_Nact" + str(
-            int(self.number_act))
-
-        if DMconfig[self.Name_DM + "_Generic"] == True:
-            self.string_os += "Gen"
-
-        if self.active == False:
-            print(self.Name_DM + ' is not activated')
-            return
-
-        self.DMconfig = DMconfig
-
-        # We need a pupil in creatingpushact_inpup() and for
-        # which in pup. THIS IS NOT THE ENTRANCE PUPIL,
-        # this is a round pupil of the same size
-        self.clearpup = Pupil(modelconfig, PupType="RoundPup", prad=self.prad)
-
-        # create the DM_pushact, surface of the DM for each individual act
-        # DM_pushact is always in the DM plane
-        self.DM_pushact = self.creatingpushact(DMconfig)
-
-        # create or load 'which actuators are in pupil'
-        self.WhichInPupil = self.id_in_pupil_actuators()
-
-        self.misregistration = DMconfig[self.Name_DM + "_misregistration"]
-        # now if we relaunch self.DM_pushact, and if misregistration = True
-        # it will be different due to misregistration
-
-    def EF_through(self,
-                   entrance_EF=1.,
-                   wavelength=None,
-                   DMphase=0.,
-                   save_all_planes_to_fits=False,
-                   dir_save_all_planes=None,
-                   **kwargs):
-        """ --------------------------------------------------
-        Propagate the electric field through the DM.
-        if z_DM = 0, then it's just a phase multiplication
-        if z_DM != 0, this is where we do the fresnel
-
-        AUTHOR : Johan Mazoyer
-
-        Parameters
-        ----------
-        entrance_EF:    2D complex array of size [self.dim_overpad_pupil, self.dim_overpad_pupil] or float scalar in which case entrance_EF is constant
-                        default is 1.
-                        Electric field in the pupil plane a the entrance of the system.
-
-        wavelength : float. Default is self.wavelength_0 the reference wavelength
-                    current wavelength in m.
-
-        DMphase : 2D array of size [self.dim_overpad_pupil, self.dim_overpad_pupil],can be complex or float scalar in which case DM_phase is constant
-                    default is 0.
-
-                    CAREFUL !! If the DM is part of a testbed. this variable name is changed
-                    to DMXXphase (DMXX: name of the DM) to avoid confusion with
-
-        save_all_planes_to_fits: Bool, default False
-                if True, save all planes to fits for debugging purposes to dir_save_all_planes
-                This can generate a lot of fits especially if in a loop so the code force you
-                to define a repository.
-
-        dir_save_all_planes : path, default None 
-                            directory to save all plane in
-                                    fits if save_all_planes_to_fits = True
-
-        Returns
-        ------
-        exit_EF : 2D array, of size [self.dim_overpad_pupil, self.dim_overpad_pupil]
-            Electric field in the pupil plane a the exit of the system
-
-
-        -------------------------------------------------- """
-
-        # call the OpticalSystem super function to check
-        # and format the variable entrance_EF
-        entrance_EF = super().EF_through(entrance_EF=entrance_EF)
-
-        if wavelength is None:
-            wavelength = self.wavelength_0
-
-        if isinstance(DMphase, (int, float, np.float)):
-            DMphase = np.full((self.dim_overpad_pupil, self.dim_overpad_pupil), np.float(DMphase))
-
-        if save_all_planes_to_fits == True:
-            name_plane = 'EF_PP_before_' + self.Name_DM + '_wl{}'.format(int(wavelength * 1e9))
-            useful.save_plane_in_fits(dir_save_all_planes, name_plane, entrance_EF)
-            name_plane = 'phase_' + self.Name_DM + '_wl{}'.format(int(wavelength * 1e9))
-            useful.save_plane_in_fits(dir_save_all_planes, name_plane, DMphase)
-
-        # if the DM is not active or if the surface is 0
-        # we save some time : the EF is not modified
-        if self.active == False or (DMphase == 0.).all():
-            return entrance_EF
-
-        if self.z_position == 0:
-            EF_after_DM = entrance_EF * self.EF_from_phase_and_ampl(phase_abb=DMphase, wavelengths=wavelength)
-
-        else:
-            EF_after_DM = self.prop_pup_to_DM_and_back(entrance_EF,
-                                                       DMphase,
-                                                       wavelength,
-                                                       save_all_planes_to_fits=save_all_planes_to_fits,
-                                                       dir_save_all_planes=dir_save_all_planes)
-
-        if save_all_planes_to_fits == True:
-            name_plane = 'EF_PP_after_' + self.Name_DM + '_wl{}'.format(int(wavelength * 1e9))
-            useful.save_plane_in_fits(dir_save_all_planes, name_plane, EF_after_DM)
-
-        return EF_after_DM
-
-    def creatingpushact(self, DMconfig):
-        """ --------------------------------------------------
-        OPD map induced in the DM plane for each actuator.
-
-        This large array is initialized at the beginning and will be use
-        to transorm a voltage into a phase for each DM. This is saved
-        in .fits to save times if the parameter have not changed
-
-        In case of "misregistration = True" we measure it once for
-        creating the interaction matrix and then once again, between
-        the matrix measrueemnt and the correction with a small mismatch
-        to simulate its effect.
-
-        AUTHOR : Axel Potier
-
-        Parameters
-        ----------
-        DMconfig : dict
-            DM configuration parameters dictionary 
-
-        Returns
-        ------
-        pushact : 3D numpy array 
-                    of size [self.number_act, self.dim_overpad_pupil, self.dim_overpad_pupil]  
-                    contains all the DM OPD map induced in the DM plane for each actuator.
-
-        
-        -------------------------------------------------- """
-        start_time = time.time()
-        Name_pushact_fits = "PushAct_" + self.Name_DM
-
-        if DMconfig[self.Name_DM + "_Generic"] == True:
-            Name_pushact_fits += "Gen"
-
-        Name_pushact_fits += "_Nact" + str(int(self.number_act)) + '_dimPP' + str(int(
-            self.dim_overpad_pupil)) + '_prad' + str(int(self.prad))
-
-        if (self.misregistration is False) and (os.path.exists(self.Model_local_dir + Name_pushact_fits +
-                                                               '.fits')):
-            pushact3d = fits.getdata(os.path.join(self.Model_local_dir, Name_pushact_fits + '.fits'))
-            print("Load " + Name_pushact_fits)
-            return pushact3d
-
-        if self.misregistration:
-            xerror = float(DMconfig[self.Name_DM + "_xerror"])
-            yerror = float(DMconfig[self.Name_DM + "_yerror"])
-            angerror = float(DMconfig[self.Name_DM + "_angerror"])
-            gausserror = float(DMconfig[self.Name_DM + "_gausserror"])
-        else:
-            xerror = 0.
-            yerror = 0.
-            angerror = 0.
-            gausserror = 0.
-
-        diam_pup_in_pix = int(2 * self.prad)
-        diam_pup_in_m = self.diam_pup_in_m
-        dim_array = self.dim_overpad_pupil
-
-        filename_actu_infl_fct = DMconfig[self.Name_DM + "_filename_actu_infl_fct"]
-
-        if DMconfig[self.Name_DM + "_Generic"] == False:
-            #Measured positions for each actuator in pixel with (0,0) = center of pupil
-            simu_grid = fits.getdata(
-                model_dir + DMconfig[self.Name_DM + "_filename_grid_actu"]) * diam_pup_in_pix + dim_array / 2
-            # the DM pitchs are read in the header
-            pitchDMX = fits.getheader(model_dir +
-                                      DMconfig[self.Name_DM + "_filename_grid_actu"])["PitchV"] * 1e-6
-            pitchDMY = fits.getheader(model_dir +
-                                      DMconfig[self.Name_DM + "_filename_grid_actu"])["PitchH"] * 1e-6
-        else:
-            # in this case we have a generic Nact1DxNact1D DM in which the pupil is centered
-            # the pitch is read in the parameter file
-            Nact1D = DMconfig[self.Name_DM + "_Nact1D"]
-            pitchDM = DMconfig[self.Name_DM + "_pitch"]
-            simu_grid = proc.generic_actuator_position(Nact1D, pitchDM, diam_pup_in_m,
-                                                       diam_pup_in_pix) + dim_array / 2
-            pitchDMX = pitchDMY = pitchDM
-
-        # Influence function and the pitch in pixels
-        actshape = fits.getdata(model_dir + filename_actu_infl_fct)
-        pitch_actshape = fits.getheader(model_dir + filename_actu_infl_fct)['PITCH']
-
-        # Scaling the influence function to the desired dimension
-        # for numerical simulation
-        resizeactshape = proc.ft_zoom_out(actshape,
-                                          (diam_pup_in_pix / diam_pup_in_m * pitchDMX / pitch_actshape,
-                                           diam_pup_in_pix / diam_pup_in_m * pitchDMY / pitch_actshape))
-
-        # make sure the actuator shape is in a squarre array of enven dimension (useful for the fft shift).
-        # We do not care exactly about the centering since we recenter the actuator just after
-        dim_even = int(np.ceil(np.max(resizeactshape.shape) / 2 + 1)) * 2
-        resizeactshape = proc.crop_or_pad_image(resizeactshape, dim_even)
-
-        # Gauss2Dfit for centering the rescaled influence function
-        Gaussian_fit_param = proc.gauss2Dfit(resizeactshape)
-        dx = Gaussian_fit_param[3]
-        dy = Gaussian_fit_param[4]
-        xycent = len(resizeactshape) / 2
-
-        # Center the actuator shape on a pixel and normalize
-        resizeactshape = proc.ft_subpixel_shift(resizeactshape, xshift=xycent - dx,
-                                                yshift=xycent - dy) / np.amax(resizeactshape)
-
-        # Put the centered influence function inside an array (self.dim_overpad_pupil x self.dim_overpad_pupil)
-        actshapeinpupil = proc.crop_or_pad_image(resizeactshape, dim_array)
-        xycenttmp = len(actshapeinpupil) / 2
-
-        # Fill an array with the influence functions of all actuators
-        pushact3d = np.zeros((simu_grid.shape[1], dim_array, dim_array))
-
-        # do the first FT only once
-        ft_actu = np.fft.fftshift(np.fft.fft2(np.fft.fftshift(actshapeinpupil), norm="ortho"))
-
-        for i in np.arange(pushact3d.shape[0]):
-
-            # Add an error on the orientation of the grid
-            if angerror != 0:
-                simu_grid[1, i] = simu_grid[1, i] * np.cos(np.radians(angerror)) - simu_grid[0, i] * np.sin(
-                    np.radians(angerror))
-                simu_grid[0, i] = simu_grid[1, i] * np.sin(np.radians(angerror)) + simu_grid[0, i] * np.cos(
-                    np.radians(angerror))
-
-            Psivector = proc.ft_subpixel_shift(ft_actu,
-                                               xshift=simu_grid[1, i] - xycenttmp + xerror * pitch_actshape,
-                                               yshift=simu_grid[0, i] - xycenttmp + yerror * pitch_actshape,
-                                               fourier=True,
-                                               norm="ortho")
-
-            if gausserror != 0:
-                # Add an error on the sizes of the influence functions
-
-                xy0 = np.unravel_index(Psivector.argmax(), Psivector.shape)
-                x, y = np.mgrid[0:dim_array, 0:dim_array]
-                xy = (x, y)
-                Psivector = proc.twoD_Gaussian(xy,
-                                               1,
-                                               1 + gausserror,
-                                               1 + gausserror,
-                                               xy0[0],
-                                               xy0[1],
-                                               0,
-                                               0,
-                                               flatten=False)
-            Psivector[np.where(Psivector < 1e-4)] = 0
-
-            pushact3d[i] = Psivector
-
-        # we exclude the actuators non active
-        pushact3d = pushact3d[self.active_actuators]
-
-        if self.misregistration is False and (
-                not os.path.exists(self.Model_local_dir + Name_pushact_fits + '.fits')):
-            fits.writeto(self.Model_local_dir + Name_pushact_fits + '.fits', pushact3d)
-            print("time for " + Name_pushact_fits + " (s):", round(time.time() - start_time))
-
-        return pushact3d
-
-    def id_in_pupil_actuators(self):
-        """ --------------------------------------------------
-        Create a vector with the index of all the actuators located in the entrance pupil
-        
-        AUTHOR: Johan Mazoyer
-
-        Parameters
-        ----------
-        cutinpupil: float
-                    minimum surface of an actuator inside the pupil to be taken into account
-                    (between 0 and 1, ratio of an actuator perfectly centered in the entrance pupil)
-
-        Returns
-        ------
-        WhichInPupil: 1D array
-                index of all the actuators located inside the pupil
-
-        
-        -------------------------------------------------- """
-        start_time = time.time()
-        Name_WhichInPup_fits = "WhichInPup_" + self.Name_DM
-
-        if self.DMconfig[self.Name_DM + "_Generic"] == True:
-            Name_WhichInPup_fits += "Gen"
-
-        Name_WhichInPup_fits += "_Nact" + str(int(self.number_act)) + '_dimPP' + str(
-            int(self.dim_overpad_pupil)) + '_prad' + str(int(self.prad)) + "_thres" + str(
-                self.WhichInPup_threshold)
-
-        if os.path.exists(self.Model_local_dir + Name_WhichInPup_fits + '.fits'):
-            print("Load " + Name_WhichInPup_fits)
-            return fits.getdata(self.Model_local_dir + Name_WhichInPup_fits + '.fits')
-
-        if self.z_position != 0:
-
-            Pup_inDMplane = proc.crop_or_pad_image(
-                prop.prop_angular_spectrum(self.clearpup.pup, self.wavelength_0, self.z_position,
-                                           self.diam_pup_in_m / 2, self.prad), self.dim_overpad_pupil)
-        else:
-            Pup_inDMplane = self.clearpup.pup
-
-        WhichInPupil = []
-        Sum_actu_with_pup = np.zeros(self.number_act)
-
-        for num_actu in np.arange(self.number_act):
-            Sum_actu_with_pup[num_actu] = np.sum(np.abs(self.DM_pushact[num_actu] * Pup_inDMplane))
-
-        Max_val = np.max(Sum_actu_with_pup)
-        for num_actu in np.arange(self.number_act):
-            if Sum_actu_with_pup[num_actu] > Max_val * self.WhichInPup_threshold:
-                WhichInPupil.append(num_actu)
-
-        WhichInPupil = np.array(WhichInPupil)
-
-        fits.writeto(self.Model_local_dir + Name_WhichInPup_fits + '.fits', WhichInPupil, overwrite=True)
-        print("time for " + Name_WhichInPup_fits + " (s):", round(time.time() - start_time))
-
-        return WhichInPupil
-
-    def prop_pup_to_DM_and_back(self,
-                                entrance_EF,
-                                phase_DM,
-                                wavelength,
-                                save_all_planes_to_fits=False,
-                                dir_save_all_planes=None):
-        """ --------------------------------------------------
-        Propagate the field towards an out-of-pupil plane ,
-        add the DM phase, and propagate to the next pupil plane
-        
-        AUTHOR : Raphaël Galicher, Johan Mazoyer
-
-        REVISION HISTORY :
-            Revision 1.1  2021-02-10 Raphaël Galicher (Initial revision)
-            Revision 2.0 2021-02-28 Johan Mazoyer (Make it more general for all DMs, put in the struc)
-
-        Parameters
-        ----------
-        pupil_wavefront : 2D array (float, double or complex)
-                    Wavefront in the pupil plane
-
-        phase_DM : 2D array
-                    Phase introduced by out of PP DM
-
-        wavelength : float
-                    wavelength in m
-
-        save_all_planes_to_fits: Bool, default False
-                if True, save all planes to fits for debugging purposes to dir_save_all_planes
-                This can generate a lot of fits especially if in a loop so the code force you
-                to define a repository.
-        
-        dir_save_all_planes : string default None. 
-                            directory to save all plane in fits if save_all_planes_to_fits = True
-
-        Returns
-        ------
-        EF_back_in_pup_plane : 2D array (complex)
-                            Wavefront in the pupil plane following the DM
-
-            
-
-        -------------------------------------------------- """
-
-        EF_inDMplane = proc.crop_or_pad_image(
-            prop.prop_angular_spectrum(entrance_EF, wavelength, self.z_position, self.diam_pup_in_m / 2.,
-                                       self.prad), self.dim_overpad_pupil)
-
-        if save_all_planes_to_fits == True:
-            name_plane = 'EF_before_DM_in_' + self.Name_DM + 'plane_wl{}'.format(int(wavelength * 1e9))
-            useful.save_plane_in_fits(dir_save_all_planes, name_plane, EF_inDMplane)
-
-        # Add DM phase at the right WL
-        EF_inDMplane_after_DM = EF_inDMplane * self.EF_from_phase_and_ampl(phase_abb=phase_DM,
-                                                                           wavelengths=wavelength)
-
-        if save_all_planes_to_fits == True:
-            name_plane = 'EF_after_DM_in_' + self.Name_DM + 'plane_wl{}'.format(int(wavelength * 1e9))
-            useful.save_plane_in_fits(dir_save_all_planes, name_plane, EF_inDMplane)
-
-        # and propagate to next pupil plane
-
-        EF_back_in_pup_plane = proc.crop_or_pad_image(
-            prop.prop_angular_spectrum(EF_inDMplane_after_DM, wavelength, -self.z_position,
-                                       self.diam_pup_in_m / 2., self.prad), self.dim_overpad_pupil)
-
-        return EF_back_in_pup_plane
-
-    def voltage_to_phase(self, actu_vect, einstein_sum=False):
-        """ --------------------------------------------------
-        Generate the phase applied on one DM for a give vector of actuator amplitude
-        We decided to do it without matrix multiplication to save time because a
-        lot of the time we have lot of zeros in it
-
-        The phase is define at the reference wl and multiply by wl_ratio in DM.EF_through
-
-        AUTHOR: Johan Mazoyer
-
-        Parameters
-        ----------
-        actu_vect : 1D array
-                    values of the amplitudes for each actuator
-        einstein_sum : boolean, default false
-                        Use numpy Einstein sum to sum the pushact[i]*actu_vect[i]
-                        gives the same results as normal sum. Seems ot be faster for unique actuator
-                        but slower for more complex phases
-
-        Returns
-        ------
-            DM_phase: 2D array
-                        phase map in the same unit as actu_vect * DM_pushact)
-        
-
-        -------------------------------------------------- """
-
-        where_non_zero_voltage = np.where(actu_vect != 0)
-
-        #opd is in nanometer
-        # DM_pushact is in opd nanometer
-        opd_to_phase = 2 * np.pi * 1e-9 / self.wavelength_0
-
-        if einstein_sum == True or len(where_non_zero_voltage[0]) < 3:
-            phase_on_DM = np.einsum('i,ijk->jk', actu_vect[where_non_zero_voltage],
-                                    self.DM_pushact[where_non_zero_voltage]) * opd_to_phase
-        else:
-            phase_on_DM = np.zeros((self.dim_overpad_pupil, self.dim_overpad_pupil))
-            for i in where_non_zero_voltage[0]:
-                phase_on_DM += self.DM_pushact[i, :, :] * actu_vect[i] * opd_to_phase
-
-        return phase_on_DM
-
-    def create_DM_basis(self, basis_type='actuator'):
-        """ --------------------------------------------------
-        Create a DM basis.
-        TODO do a zernike basis ?
-
-        AUTHOR: Johan Mazoyer
-
-        Parameters
-        ----------
-        basis_type: string, default 'actuator' 
-            the type of basis. 'fourier' or 'actuator'
-
-        Returns
-        ------
-        basis: 2d numpy array 
-            basis [Size basis, Number of active act in the DM]
-
-
-        -------------------------------------------------- """
-        if basis_type == 'actuator':
-            # no need to remove the inactive actuators,
-            # they are already removed in pushact
-            basis_size = len(self.WhichInPupil)
-            basis = np.zeros((basis_size, self.number_act))
-            for i in range(basis_size):
-                basis[i][self.WhichInPupil[i]] = 1
-
-        elif basis_type == 'fourier':
-            start_time = time.time()
-            # activeact = [value for value in self.active_actuators]
-
-            sqrtnbract = int(np.sqrt(self.total_act))
-            Name_FourrierBasis_fits = "Fourier_basis_" + self.Name_DM + '_prad' + str(
-                self.prad) + '_nact' + str(sqrtnbract) + 'x' + str(sqrtnbract)
-
-            cossinbasis = phase_ampl.sine_cosine_basis(sqrtnbract)
-
-            basis_size = cossinbasis.shape[0]
-            basis = np.zeros((basis_size, self.number_act))
-
-            for i in range(basis_size):
-                vec = cossinbasis[i].flatten()[self.active_actuators]
-                basis[i] = vec
-
-            start_time = time.time()
-            # This is a very time consuming part of the code.
-            # from N voltage vectors with the sine and cosine value, we go N times through the
-            # voltage_to_phase functions. For this reason we save the Fourrier base 2D phases on each DMs
-            # in a specific .fits file
-            if not os.path.exists(self.Model_local_dir + Name_FourrierBasis_fits + '.fits'):
-                phasesFourrier = np.zeros((basis_size, self.dim_overpad_pupil, self.dim_overpad_pupil))
-                print("Start " + Name_FourrierBasis_fits)
-                for i in range(basis_size):
-                    phasesFourrier[i] = self.voltage_to_phase(basis[i])
-                    if i % 10:
-                        useful._progress(i, basis_size, status='')
-                fits.writeto(self.Model_local_dir + Name_FourrierBasis_fits + '.fits', phasesFourrier)
-            print("time for " + Name_FourrierBasis_fits, time.time() - start_time)
-
-        else:
-            raise Exception(basis_type + " is is not a valid basis_type")
-
-        return basis
-
-
-class Testbed(OpticalSystem):
-    """ --------------------------------------------------
-    
-    Initialize and describe the behavior of a testbed.
-    This is a particular subclass of Optical System, because we do not know what is inside
-    It can only be initialized by giving a list of Optical Systems and it will create a
-    "testbed" with contains all the Optical Systems and associated EF_through functions and
-    correct normlaization
-
-    AUTHOR : Johan Mazoyer
-
-    -------------------------------------------------- """
-
-    def __init__(self, list_os, list_os_names):
-        """ --------------------------------------------------
-        This function allows you to concatenate OpticalSystem objects to create a testbed:
-        parameter:
-            list_os:        list of OpticalSystem instances
-                            all the systems must have been defined with
-                            the same modelconfig or it will send an error.
-                            The list order is form the first optics system to the last in the
-                            path of the light (so usually from entrance pupil to Lyot pupil)
-
-            list_os_names:  list of string of the same size as list_os 
-                            Name of the optical systems. 
-                            Then can then be accessed inside the Testbed object by os_#i = Testbed.list_os_names[i]
-
-        Returns
-        ------
-            testbed : an optical system which is the concatenation of all the optical systems
-
-
-
-        -------------------------------------------------- """
-        if len(list_os) != len(list_os_names):
-            print("")
-            raise Exception("list of systems and list of names need to be of the same size")
-
-        # Initialize the OpticalSystem class and inherit properties
-        super().__init__(list_os[0].modelconfig)
-
-        init_string = self.string_os
-
-        # Initialize the EF_through_function
-        self.EF_through = super().EF_through
-
-        self.number_DMs = 0
-        self.number_act = 0
-        self.name_of_DMs = list()
-
-        # this is the collection of all the possible keywords that can be used in
-        # practice in the final testbed.EF_through, so that can be used in
-        # all the EF_through functions
-        known_keywords = list()
-
-        # we store the name of all the sub systems
-        self.subsystems = list_os_names
-
-        # we concatenate the Optical Element starting by the end
-        for num_optical_sys in range(len(list_os)):
-
-            # we first check that all variables in the list are optical systems
-            # defined the same way.
-            if not isinstance(list_os[num_optical_sys], OpticalSystem):
-                raise Exception("list_os[" + str(num_optical_sys) + "] is not an optical system")
-
-            if list_os[num_optical_sys].modelconfig != self.modelconfig:
-                print("")
-                raise Exception("All optical systems need to be defined with the same initial modelconfig!")
-
-            # if the os is a DM we increase the number of DM counter and
-            # store the number of act and its name
-
-            for params in inspect.signature(list_os[num_optical_sys].EF_through).parameters:
-                known_keywords.append(params)
-
-            if isinstance(list_os[num_optical_sys], DeformableMirror):
-
-                #this function is to replace the DMphase variable by a XXphase variable
-                # where XX is the name of the DM
-                list_os[num_optical_sys].EF_through = _swap_DMphase_name(
-                    list_os[num_optical_sys].EF_through, list_os_names[num_optical_sys] + "phase")
-                known_keywords.append(list_os_names[num_optical_sys] + "phase")
-
-                if list_os[num_optical_sys].active == False:
-                    # if the Dm is not active, we just add it to the testbed model
-                    # but not to the EF_through function
-                    vars(self)[list_os_names[num_optical_sys]] = list_os[num_optical_sys]
-                    continue
-
-                self.number_DMs += 1
-                self.number_act += list_os[num_optical_sys].number_act
-                self.name_of_DMs.append(list_os_names[num_optical_sys])
-
-            # concatenation of the EF_through functions
-            self.EF_through = _concat_fun(list_os[num_optical_sys].EF_through, self.EF_through)
-
-            # we add all systems to the Optical System so that they can be accessed
-            vars(self)[list_os_names[num_optical_sys]] = list_os[num_optical_sys]
-
-            self.string_os += list_os[num_optical_sys].string_os.replace(init_string, '')
-
-        # in case there is no coronagraph in the system, we still add
-        # noFPM so that it does not break when we run transmission and max_sum_PSFs
-        # which pass this keyword by default
-        known_keywords.append('noFPM')
-        known_keywords.append('photon_noise')
-        known_keywords.append('nb_photons')
-        known_keywords.append('in_contrast')
-
-        # we remove doubloons
-        # known_keywords = list(set(known_keywords))
-        known_keywords = list(dict.fromkeys(known_keywords))
-
-        # We remove arguments we know are wrong
-        if 'DMphase' in known_keywords:
-            known_keywords.remove('DMphase')
-        if self.number_DMs > 0:
-            # there is at least a DM, we add voltage_vector as an authorize kw
-            known_keywords.append('voltage_vector')
-            self.EF_through = _control_testbed_with_voltages(self, self.EF_through)
-
-        # to avoid mis-use we only use specific keywords.
-        known_keywords.remove('kwargs')
-
-        self.EF_through = _clean_EF_through(self.EF_through, known_keywords)
-
-        #initialize the max and sum of PSFs for the normalization to contrast
-        self.measure_normalization()
-
-    def voltage_to_phases(self, actu_vect, einstein_sum=False):
-        """ --------------------------------------------------
-        Generate the phase applied on each DMs of the testbed from a given vector of
-        actuator amplitude. I split theactu_vect and  then for each DM, it uses
-        DM.voltage_to_phase (no s)
-
-        Parameters
-        ----------
-        actu_vect : float or 1D array of size testbed.number_act
-                    values of the amplitudes for each actuator and each DM
-        einstein_sum : boolean. default false
-                        Use numpy Einstein sum to sum the pushact[i]*actu_vect[i]
-                        gives the same results as normal sum. Seems ot be faster for unique actuator
-                        but slower for more complex phases
-
-        Returns
-        ------
-            3D array of size [testbed.number_DMs, testbed.dim_overpad_pupil,testbed.dim_overpad_pupil]
-            phase maps for each DMs by order of light path in the same unit as actu_vect * DM_pushact
-
-        AUTHOR : Johan Mazoyer
-        -------------------------------------------------- """
-        DMphases = np.zeros((self.number_DMs, self.dim_overpad_pupil, self.dim_overpad_pupil))
-        indice_acum_number_act = 0
-
-        if isinstance(actu_vect, (int, float)):
-            return np.zeros(self.number_DMs) + float(actu_vect)
-
-        if len(actu_vect) != self.number_act:
-            raise Exception("voltage vector must be 0 or array of dimension testbed.number_act," +
-                            "sum of all DM.number_act")
-
-        for i, DM_name in enumerate(self.name_of_DMs):
-
-            DM = vars(self)[DM_name]  # type: DeformableMirror
-            actu_vect_DM = actu_vect[indice_acum_number_act:indice_acum_number_act + DM.number_act]
-            DMphases[i] = DM.voltage_to_phase(actu_vect_DM, einstein_sum=einstein_sum)
-
-            indice_acum_number_act += DM.number_act
-        # useful.quickfits(DMphases, dir="/Users/jmazoyer/Desktop/DM_phase/")
-
-        return DMphases
-
-    def basis_vector_to_act_vector(self, vector_basis_voltage):
-        """ --------------------------------------------------
-        transform a vector of voltages on the mode of a basis in a  vector of
-        voltages of the actuators of the DMs of the system
-
-        Parameters
-        ----------
-        vector_basis_voltage: 1D-array real : 
-                        vector of voltages of size (total(basisDM sizes)) on the mode of the basis for all
-                        DMs by order of the light path
-
-        Returns
-        ------
-        vector_actuator_voltage: 1D-array real : 
-                        vector of base coefficients for all actuators of the DMs by order of the light path
-                        size (total(DM actuators))
-        
-
-        -------------------------------------------------- """
-
-        indice_acum_basis_size = 0
-        indice_acum_number_act = 0
-
-        vector_actuator_voltage = np.zeros(self.number_act)
-        for DM_name in self.name_of_DMs:
-
-            # we access each DM object individually
-            DM = vars(self)[DM_name]  # type: DeformableMirror
-
-            # we extract the voltages for this one
-            # this voltages are in the DM basis
-            vector_basis_voltage_for_DM = vector_basis_voltage[indice_acum_basis_size:indice_acum_basis_size +
-                                                               DM.basis_size]
-
-            # we change to actuator basis
-            vector_actu_voltage_for_DM = np.dot(np.transpose(DM.basis), vector_basis_voltage_for_DM)
-
-            # we recreate a voltages vector, but for each actuator
-            vector_actuator_voltage[indice_acum_number_act:indice_acum_number_act +
-                                    DM.number_act] = vector_actu_voltage_for_DM
-
-            indice_acum_basis_size += DM.basis_size
-            indice_acum_number_act += DM.number_act
-
-        return vector_actuator_voltage
-
-
-# Some internal functions to properly concatenate the EF_through functions
-def _swap_DMphase_name(DM_EF_through_function, name_var):
-    """ --------------------------------------------------
-   A function to rename the DMphase parameter to another name (usually DMXXphase)
-        
-    AUTHOR : Johan Mazoyer
-
-    Parameters:
-    ------
-        DM_EF_through_function : function
-            the function of which we want to change the params
-        name_var : string 
-            the name of the  new name variable
-
-    Returns
-    ------
-        the_new_function: function
-            with name_var as a param
-
-    
-    -------------------------------------------------- """
-
-    def wrapper(**kwargs):
-
-        if name_var not in kwargs.keys():
-            kwargs[name_var] = 0.
-        new_kwargs = copy.copy(kwargs)
-
-        new_kwargs['DMphase'] = kwargs[name_var]
-
-        return DM_EF_through_function(**new_kwargs)
-
-    return wrapper
-
-
-def _concat_fun(outer_EF_through_fun, inner_EF_through_fun):
-    """ --------------------------------------------------
-    A very small function to concatenate 2 functions
-    AUTHOR : Johan Mazoyer
-
-    Parameters:
-    ------
-        outer_fun: function
-                x -> outer_fun(x)
-        inner_fun: function 
-                x -> inner_fun(x)
-
-    Returns
-        ------
-        the concatenated function: function
-                x -> outer_fun(inner_fun(x))
-
-
-
-    -------------------------------------------------- """
-
-    def new_EF_through_fun(**kwargs):
-
-        new_kwargs_outer = copy.copy(kwargs)
-        del new_kwargs_outer['entrance_EF']
-
-        return outer_EF_through_fun(entrance_EF=inner_EF_through_fun(**kwargs), **new_kwargs_outer)
-
-    return new_EF_through_fun
-
-
-def _clean_EF_through(testbed_EF_through, known_keywords):
-    """ --------------------------------------------------
-    a functions to check that we do not set unknown keyword in
-    the testbed EF through function. Maybe not necessary.
-
-    AUTHOR : Johan Mazoyer
-
-    Parameters:
-    ------
-         testbed_EF_through: function
-         known_keywords: list of strings of known keywords
-
-    Returns
-    ------
-        cleaned_testbed_EF_through: function
-            a function where only known keywords are allowed
-        
-
-    -------------------------------------------------- """
-
-    def wrapper(**kwargs):
-        for passed_arg in kwargs.keys():
-            if passed_arg == 'DMphase':
-                raise Exception('DMphase is an ambiguous argument if you have several DMs.' +
-                                ' Please use XXphase with XX = nameDM')
-            if passed_arg not in known_keywords:
-                raise Exception(passed_arg + 'is not a EF_through valid argument. Valid args are ' +
-                                str(known_keywords))
-
-        return testbed_EF_through(**kwargs)
-
-    return wrapper
-
-
-def _control_testbed_with_voltages(testbed: Testbed, testbed_EF_through):
-    """ --------------------------------------------------
-    A function to go from a testbed_EF_through with several DMXX_phase
-    parameters (one for each DM), to a testbed_EF_through with a unique
-    voltage_vector parameter of size testbed.number_act (or a single float, like 0.)
-
-    the problem with DMXX_phase parameters is that it cannot be automated since it requires
-    to know the name/number of the DMs in advance.
-
-    DMXX_phase parameters can still be used, but are overridden by voltage_vector parameter
-    if present.
-
-    AUTHOR : Johan Mazoyer
-
-    Parameters:
-    ------
-        DM_EF_through_function : function
-                the function of which we want to change the params
-        name_var : string 
-                the name of the  new name variable
-
-    Returns
-    ------
-        the_new_function: function
-                with name_var as a param
-
-
-    -------------------------------------------------- """
-
-    def wrapper(**kwargs):
-        if 'voltage_vector' in kwargs:
-            voltage_vector = kwargs['voltage_vector']
-            DM_phase = testbed.voltage_to_phases(voltage_vector)
-            for i, DM_name in enumerate(testbed.name_of_DMs):
-                name_phase = DM_name + "phase"
-                kwargs[name_phase] = DM_phase[i]
-
-        return testbed_EF_through(**kwargs)
-
-    return wrapper
